requirements.txt
asgiref==3.3.1
Django==2.2.10
mysqlclient==2.0.3
Pillow==8.2.0
python-barcode==0.13.1
pytz==2021.1
sqlparse==0.4.1
